Unable to connect category with class.
I have been trying different angles with this and I am still coming up with the same issue. In the spec/models/category_spec.rb file. It is unable to connect the class Category from the models. I am guessing that this might be a naming convention thing but I am not sure.

next:
-escalated to jack